Renamed a few things here and there to break it up a bit.
