+++ src/dolomite/__init__.py
@@ -19,4 +19,4 @@ from .stage_object import stage_object
 from .stage_data_frame import stage_data_frame
 from .write_metadata import write_metadata
 from .load_object import *
-from .load_data_frame import load_data_frame_csv
+from .load_csv_data_frame import load_csv_data_frame
